requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.116.2
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
h11==0.16.0
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.9
markupsafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
polars==1.33.1
psutil==7.0.0
pydantic==2.11.9
pydantic-core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
# torch==2.8.0+cu126
# torchvision==0.23.0+cu126
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.201
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
